pyproject.toml
[tool.poetry]
name = "di-identity-vocab"
version = "1.0.0"
description = "GOV.UK identity vocabulary"
authors = ["Your Name <you@example.com>"]
license = "MIT Copyright (c) 2022 Crown Copyright (Government Digital Service)"

[tool.poetry.dependencies]
python = "^3.10"
linkml = {path = "./linkml", develop = true}
mkdocs = "^1.3.1"
mkdocs-tech-docs-template = "^0.1.2"
mkdocs-mermaid2-plugin = "^1.1.1"
mkdocs-material = "^9.4.6"
ruamel-yaml = "^0.18.2"

[tool.poetry.dev-dependencies]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
